+++ WalkTheMaze-Object-oriented/BizzKit/Upload.cpp
@@ -1,7 +1,7 @@
 #include "Upload.h"
 #include "ServicesHandler.h"
 
-int Upload::uploadProgress(string username, int score, GLfloat pos_x, GLfloat pos_y)
+int Upload::uploadProgress(string username, int score, int pos_x, int pos_y)
 {
 	ServicesHandler *service = new ServicesHandler;
 	int res = service->uploadRoute(username, score, pos_x, pos_y);
