+++ WalkTheMaze-Object-oriented/BizzKit/Upload.h
@@ -14,8 +14,8 @@ using namespace std;
 class Upload{
 
 public:
-	static int uploadProgress(string username, int score, GLfloat pos_x, GLfloat pos_y);
-	static int saveProgress(string user, int score, GLfloat pos_x, GLfloat pos_y);
+	static int uploadProgress(string username, int score, int pos_x, int pos_y);
+	static int saveProgress(string user, int score, int pos_x, int pos_y);
 };
 
 #endif
